[JENKINS-38805_User_can_run_parametrised_pipeline] refactor runButton usage to now to use it as you would runButton but with parameter support
